backend/requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pillow==9.5.0
scikit-learn==1.2.2
# torch==2.0.1
# torchvision==0.15.2
tqdm==4.65.0
ipykernel==6.23.1
urllib3==1.26.7
boto3==1.18.63
